Secured against non owners

+++ room_view/index.php
@@ -15,6 +15,10 @@ if (!isset($action)) {
 switch ($action) {
     case "show_details":
         $room_id = filter_input(INPUT_GET, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $room = get_room_by_id($room_id);
         $questions = get_questions_by_room($room_id);
         $machine_types = get_all_machine_types();
@@ -22,6 +26,10 @@ switch ($action) {
         break;
     case "create":
         $room_id = filter_input(INPUT_GET, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         if (get_room_by_id($room_id)["room_code"] == null) {
             $used_hashes = get_all_hashes();
 
@@ -38,6 +46,10 @@ switch ($action) {
     case "add_case":
         $question_id = filter_input(INPUT_POST, "question_id");
         $room_id = filter_input(INPUT_POST, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $test_input = filter_input(INPUT_POST, "test_input");
         $pass = filter_input(INPUT_POST, "pass") ? "1": "0";
         add_test_cases($question_id, $test_input, $pass);
@@ -45,12 +57,20 @@ switch ($action) {
         break;
     case "get_test_cases":
         $question_id = filter_input(INPUT_GET, "question_id");
+        if (!(is_question_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $cases = get_test_cases($question_id);
         header("Content-Type: application/json");
         echo json_encode(array("cases" => $cases));
         break;
     case "add_question":
         $room_id = filter_input(INPUT_POST, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $question_name = filter_input(INPUT_POST, "question_name");
         $machine_type = filter_input(INPUT_POST, "machine_type");
         add_question($room_id, $question_name, $machine_type);
@@ -58,12 +78,20 @@ switch ($action) {
         break;
     case "delete_question":
         $room_id = filter_input(INPUT_GET, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $question_id = filter_input(INPUT_GET, "question_id");
         delete_question($question_id);
         header("Location: ./index.php?room_id=" . $room_id);
         break;
     case "show_users":
         $room_id = filter_input(INPUT_GET, "room_id");
+        if (!(is_owner($room_id, $current_user["user_id"]))) {
+            header("Location: ../rooms/index.php");
+            exit();
+        }
         $room = get_room_by_id($room_id);
         $user_list = get_users_by_room($room_id);
         $questions = get_questions_by_room($room_id);
